pyproject.toml
[project]
name = "cs-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "langchain>=1.1.0",
    "langchain-anthropic>=1.2.0",
    "langchain-openai>=1.1.0",
    "langfuse>=3.10.5",
    "langgraph>=1.0.4",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.8",
]

[tool]

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]
